change behaviour of list parsing option
